+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Basic help entry.
 - Color Mappings can be added in scripts.
+- Add enable/disable toggle in toolbar.
 
 ## [1] - 2018-07-18